requirements.txt
flask==3.0.3
python-dotenv==1.0.1
bcrypt==4.1.3
requests==2.31.0
reportlab==4.2.2
gTTS==2.5.1
